Benchmarking integrals code. Integrals and Gradients work now for no permutational symmetry.  The standalone validation code N4K uses Compare.nb to compare the HGP routine with the MD integral routine in MondoMods/McMurchie.F90.  The standalone validation code dN4K uses finite differences of raw integrals to validate the gradients code.  HRR.nb is a notebook I used in debuging the gradients.  It has the (00|00) integral, which can be analytically differentiated then evaluated to check intermediate steps in the VRR and HRR.
Of course, we will want to go back to using permutational symmetry,
however this serves as a good reference point.  Note that I havent
changed any of the subroutine interfaces etc, so they can be used
as is.  There may be some redundancy in the ShellPairsStruct now
with ONXDataType.  I think that the datatypes need to get cleaned
a bit more here and retained in ShellPairsStruct.  Its also a good
idea to maybe clean (for permutational symmetry) the N4K and dN4K
routines so that we have a demo integrals driver/checker.  This should
help us when we do N-scaling MP2...
VS: ----------------------------------------------------------------------