Removed dependence on MondoLogger.h
The regular expression syntax is slightly different for different platforms
(linux and OS X for instance). So instead of using macros to call the MondoLog
function and then apply a sed expression to fix concatenated lines, we now call
MondoLog directly without the need for macros and the sed command. This
hopefully makes it easier to build mondo on different platforms.