Fixed distcheck.
distcheck uses VPATH and some other trickery, which means that we can not
write the preprocessed fortran files into the same directory the sources came
from. This directory is write protected during distcheck. We now use the
target name (i.e. either .o or .lo) to compute the filename for the
preprocessed file. The directory the target goes into clearly needs to be
writable.

We also can not use relative paths for includes (-I or -L) since we can not
know where the sources or the .mod files will be relative to where we are
building (--srcdir).