This looks like what we thought was a "DIIS" error. To be sure, DIIS was buggy (the posdef bug was serious), but this bug is DenMatMethods was perhaps as serious, and appeared to point at DIIS still.
Basically, the fancy (WTF is going on there anyway??)
renomralization of the density matrix trace was
throwing a discontinuity into the TR-BOMD.  This
appeared to be more severe  than just a simple
re-scaling.  IMO, this routine (NormTrace) should be
viewed with extreme suspicion with argument
"order==1".  order=0 is probably ok, but not clear
that we really need it, even under extreme thresholding.

These are issues that are, I think, well outside
Manny's matrix analysis, and are best left unprovoked
for now.