Fixing MD restarts.
In P2Use:DMTRBO we calculate the drift of the auxiliary density matrix from
the "true" ground state density matrix. This measure requires that we load the
density matrix. On a restart, the density matrix is not available yet for the
very first call to P2Use. I included a check now that make this work.