Added check for naming convention in lapack.
In lapack/install we need to link either the internal etime() or an external
etime(). In the external case, we need to make sure to add an underscore in
some cases (some compiler combinations) and not in other cases. I have added a
test to configure to automatically do this correctly.