Changed makefile dependencies.
The Dependencies on MondoMods are only important in the case of static
linking. In this case we want to make sure that we relink the binaries with
the new MondoMods library. In the case of dynamic linking however, most of the
time relinking is unecessary. It only needs to be done when the public
interface of MondoMods changes.