Added configure option to make names case sensitive in fortran.
This option is for debugging. Usually fortran is case insensitive, but for
internal source code consistency it helps if symbols are always spelled the
same throughout the whole program. This should include the case.

I turned off the passing of compiler flags to hdf5. Hdf5's configure script
can figure this out itself.