Made RecycleHDF and CleanScratch input options.
The number of geometries kept in hdf is now an input file option and not a
compile time option. Specify this with

RecycleHDF = 20

in the input file. I have also added a logical option whether to clean out the
scratch directory or not. Set this with

CleanScratch = (yes|true|no|false)

The default is to clean scratch.