Removed printing to geometry file.
For restarts, we used to print all previous geometries to GFile (the geometry
file). This doesn't work well right now. We print a lot of additional
information in the comment line, such as simulation time, kinetic, potential,
and total energies, etc. while running an MD. To properly print coordinates
from before the restart, all of this information contained in the comment line
has to be extracted from hdf and properly formatted. Otherwise parsing of the
GFile is going to be tricky.