Added script to set localversion.
I removed the automated check of localversion from the makefile since it
depends on being in a git repository. Since localversion is not updated
automatically now anymore, we need to run fix_localversion.sh to update it by
hand. The makefile target dist still updates localversion automatically, since
we are assuming that one would only try to make this target if developing in a
git repository.