Fixed uninitialized variable problem in InCoords.
Some of the BondDATA were not properly initialized and caused a possible
memory problem. The number of bonds in BondDATA%N is used in
SetEQ_BONDDATA_EQ_BONDDATA to decide whether to copy the old BondDATA object
to the new one and expand its size. If the BondDATA object is not initialized
then BondDATA%N has a possible non-zero value which leads to memory access
problems. I have added proper initialization calls to avoid this.