Scratch directory now has a unique random name.
In larger MD, NEB, or optimization calculations we generate a lot of files in
scratch. If scratch is not changed between runs, we tend to fill up scratch
with a _very_ large number of files, which usually overwhelms the filesystem,
or at least makes it respond very slowly. I have changed FreeON's behavior
such that it uses MONDO_SCRATCH (either set in the environment or at configure
time) as a base directory to create the real scratch directory underneath it.
The scratch directory's name is randomly chosen and guaranteed to be unique.
It is printed to the out file so that one can find all of the necessary files
for a restart.