Cleaned up MemMan and introduced Initialize interface.
When AllocQ'ing on a newly allocated variable we rely on the fact that the
field Alloc is never accidentally set to ALLOCATED_TRUE. Of course there is
not guarantee that this is always true, however the assumption seems to have a
high chance of being true. When compiling the code with
--disable-optimizations and gcc I sometimes run into the problem that a newly
allocated variable is already flagged as allocated. I added an "Initialize"
interface to MemMan which explicitly sets the Alloc field to ALLOCATED_FALSE.
Since that is much cleaner and guaranteed to work, we should start using this
interface more.