Added reactant and product energy options for NEB.
In the NEB we don't calculate anything for the endpoints and only use their
coordinates. In order to calculate the NEB springs we need to know their
energies though. I added two new input file options, NEBReactantEnergy and
NEBProductEnergy, so that these energies can be put in through the input file.