Changed stepsize cutoff to RMSd.
The input file parameter ConjugateGradientMaxMove controls the maximum
stepsize taken in the quasi-Newton step during the optimization. I changed how
that stepsize is calculated to an RMSd value so that it is independent of
system size. This allows us to use the same value for MaxMove in NEB and
single structure optimization.