Adding basic template of regression tests.
The template uses the same fork/invoke technology as the front end. It tries
to demonstrate how to set up a regression test suite in FreeON such that one
can "make check" in the builddir without having to "make install". This is
currently not possible because we hardcode the $bindir into the front end.