Instituting executive decision to stop supporting density matrix archival in the HDF file.
Instituting executive decision to stop supporting density matrix archival in
the HDF file, because it makes parallelism almost impossible in its current
implementation.  I'm sure, with parallel HDF5 it will be possible, but even
then, I'm not sure thats a good paradigm.  Probably, it would be better to
have failsafe recomputation of the DM in the case that it is lost. Since the
restart is quite broken, this is probably not a #1 priority right now.