Adding simple example of how to use MPI_Comm_spawn().
This is basically what we do in Overlay.f90:Invoke and MondoMPI.f90:FiniMPI.
Currently we see that the master process consumes a considerable amount of CPU
resources because it is polling the MPI_Barrier(). This is not what we want,
especially since we know that our slave processes have rather long runtimes.
There does not appear to be a direct way to reduce the polling frequency in
MPI_Barrier().