CPU consumption of front-end copies.
Currently (that is with openmpi-1.3.4), the rank > 0 front-end copies each
cause a load of 1 on each CPU. They do that because the MPI_Barrier()
statement uses a high polling frequency which consumes CPU cycles. The code is
therefore currently not really that usable in parallel, because the unused
rank > 0 front-end copies take away close to half of the available nodes.