Fixed bug with parallel clones.
In some programs we check for the existence of the Lowdin or the AInv matrix.
This was done only in rank 0 and then broadcast to the other ranks in case of
PARALLEL. I don't see the advantage of limiting the INQUIRE() call to rank 0.
The INQUIRE() should be very cheap and INQUIRE()s of different ranks should be
cached by the NFS server, which should mean that we won't get any performance
advantage by using MPI_Bcast().

But please, correct me if I am wrong.