Changed default compiler flags for Intel.
The -ipo flag, which is turned on with -fast on x86 and x86_64, does not work
well together with lapack/install. The symbols are thrown out of the object
files and the linker can't link them. I have replaced -fast with what it
expands to on linux x86/x86_64: -O3 -no-prec-div -xHost.