Added example for locking with MPI.
We currently hold the hdf file open in all clones. This can only work if
everybody who has the file open gets notified that somebody else changed the
file. Some filesystems do not do that, notably NFS as a stateless filesystem
does not tell clients what other clients did while the file is open. It
implements only a "close to open cache coherent" approach. We need to make
sure that only one clone at a time has the hdf file open, and that it is
closed at all other times. The locking mechanism described in this example can
achieve that.