Last commit didn't quite fix it for NFS.
I changed QCTC so that the hdf file is closed right after StartUp() and is
only openend on a Get() or a Put(). Get() and Put() are protected by exclusive
locks. I also added a 2.5 second delay after the Put() because NFS seems to
have a certain lag when it comes to figuring out that it should check whether
a file has changed or not. QCTC right now is really hackish, but it seems to
not blow up anymore for 16 clones on 16 nodes mounting one shared NFS
directory.