Reintroduced the Cartesian communicator.
Without the Cartesian communicator in StartUp() parallel clones is broken.
MakeS for example produces identical overlap matrices along the transition
path for some reason. We might want to leave the communicator business as it
was anyway (for spatial parallelism). I will have to verify that the random
hangs in the backend we saw with Cartesian communicators are fixed now and
were due to OpenMPI. We are currently testing with OpenMPI 1.4.1 which seems
to work great so far.