Changed BCSR fileformat.
Reading a BCSR file from an external application is currently not easily
possibhle, since the file is missing some important information on the basis
set size. This information is stored in the hdf file, but not in the BCSR
file. I added an compile-time option (--enable-standalone-BCSR) that adds the
missing information to the BCSR file.